RAT_Client/requirements.txt
pynput
pycryptodome
requests
Pillow
